packages/cli/templates/auth/requirements.txt
# HostKit Auth Service Dependencies
# Generated by HostKit - do not edit manually

# FastAPI framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0

# Validation and serialization
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# HTTP client (for OAuth)
httpx>=0.26.0,<1.0.0

# Utilities
python-multipart>=0.0.6,<1.0.0
